chore(security): pin Hubble TLS floor to 1.2; drop redundant shift conversion
Two gosec findings surfaced by golangci-lint v2.6.2 (not yet flagged by the
CI-pinned linter version):

- G402: the Hubble Relay TLS dial used a zero-value tls.Config, which
  accepts TLS 1.0/1.1. Extracted tlsConfig() pinning MinVersion to TLS 1.2
  (+ TestTLSConfigMinVersion).
- G115: retryDelay's `baseBackoff << uint(attempt-1)` tripped the
  int->uint overflow rule. The surrounding guards already bound attempt-1
  to [0,62] and signed shift counts are legal since Go 1.13, so the
  conversion is simply dropped — behavior unchanged, existing table tests
  cover it.

golangci-lint run ./... is now clean repo-wide on v2.6.2.

+++ internal/httpx/retry.go
@@ -83,7 +83,9 @@ func retryDelay(attempt int, resp *http.Response) time.Duration {
 	if attempt-1 >= 63 {
 		return maxDelay
 	}
-	d := baseBackoff << uint(attempt-1)
+	// Signed shift counts are legal since Go 1.13; the guards above bound
+	// attempt-1 to [0,62], so no uint conversion (and no gosec G115) is needed.
+	d := baseBackoff << (attempt - 1)
 	if d <= 0 || d > maxDelay {
 		return maxDelay
 	}
